+++ CHANGELOG.md
@@ -15,6 +15,8 @@ bodies of source code, with links to view them inside helpful.
 
 Fixed a crash on keymaps where keys are bound to anonymous functions.
 
+Improved performance (PR by @nickdrozd).
+
 # v0.11
 
 Further work on syntax highlighting performance for large code